[dynamo] Refine `LocalSource.cell_or_freevar` to `LocalSource.is_input`
The `cell_or_freevar` was added in #106403 to help us ensure
Dynamo-export only allows graph input that depends on the frame input
(rather than a captured cell, for instance).

However, when taken literally, the `cell_or_freevar` condition is
actually not accurate, because for frame inputs that are also cells
(i.e., captured by some inner function), we actually set the
`cell_or_freevar` flag to false. This makes sense, because otherwise the
existing implementation would prevent Dynamo-export to add any of these
inputs to the graph.

To help with reasoning, this patch refines the `cell_or_freevar` flag to
what we really want to check -- `is_input`, and updates the relevant use
sites.

ghstack-source-id: a0a13edbbc13e6c3e0775307c0bd0a3f85a88571
Pull Request resolved: https://github.com/pytorch/pytorch/pull/140151