[dynamo] Add a `DynamoFrameType` type above Python frame object
This patch introduces a `DynamoFrameType` to serve as a layer between
Dynamo and different versions of Python frame object. In
`DynamoFrameType`, we only register attributes Dynamo cares about (e.g.,
`f_code`, `f_locals`, etc.

This will be helpful when it comes to adding new attributes to this
`DynamoFrameType`, or dealing with Python version changes.

ghstack-source-id: 37a4d350b25d257a469e7dbea70304264088da59
Pull Request resolved: https://github.com/pytorch/pytorch/pull/140330