[dynamo] Identify pre-existing captured cells by cell id rather than content id
In `match_nested_cell`, Dynamo tried to identify pre-existing captured
cells by `(cell_name, id(cell_contents))`. This works in most cases, but
as the test added in this patch shows, it's not a complete solution.

This patch changes `match_nested_cell` to
`lookup_variable_for_captured_cell`, and does the lookup based on id of
cell objects, not their contents. This requires plumbing a tuple of
captured cell objects from different CPython versions all the way to
`InstructionTranslator.__init__`, where we store a mapping from the ids
of these cell objects, and use it later in
`UserFunctionVariable.bind_args` to look up the VT for these unboxed
cells.

ghstack-source-id: 93c50fa236a88f30db7b20803311fe4dbaf0351e
Pull Request resolved: https://github.com/pytorch/pytorch/pull/140436